+++ lib/alf/command/show.rb
@@ -16,6 +16,12 @@ module Alf
                "Enable/disable pretty print best effort") do |val|
           @pretty = val
         end
+
+        @ff = nil
+        opt.on("--ff=FORMAT", 
+               "Specify the floating point format") do |val|
+          @ff = val
+        end
       end
         
       def run(argv, requester)
@@ -25,6 +31,7 @@ module Alf
 
         # Set options on the requester
         requester.pretty = @pretty unless @pretty.nil?
+        requester.rendering_options[:float_format] = @ff unless @ff.nil?
         requester.renderer_class = (@renderer_class || 
                                     requester.renderer_class || 
                                     Text::Renderer)